Show time between GCDs in the list of weaves
oGCDs are weaved between two GCDs (or the start of the fight and a
trailing GCD). The weaving module has been updated to track both the
leading and trailing GCDs so that the time between them can be
displayed to the user. This helps users determine if the weaving was
severe or not.

The GCD information can be read by the individual job modules, though
only the BLM and SCH modules actually seemed to care which GCDs were
used. They have been updated.