Updated Statuses and Actions
I updated the data for Bard to include the new abilities in Shadowbringers, remove the abilities no longer in the game, and remove the statuses Bard no longer applies.  Also, because Straight Shot doesn't apply a crit chance buff, I removed the references to it, and removed the module from loading.