Update Barrage Module
They changed Empyreal Arrow to an ability, so you can no longer use Barrage on it.  Instead, they made it so it proc's straighter shot, so you should always use it on Refulgent Arrow.  This module should probably be updated to check to see if a Bard used Barrage while they had a Straighter Shot proc.

For now, I have removed all references to Empyreal Arrow in this module.