Updated Stats and Fixed Snapshot functionality
I updated the Level Mod and Base Substat to 80 from 70 which is self-explanitory.

I removed arcanum and the spear card from being used as it no longer gives crit, and updated the other two crit buffs to their new values.  When dancers statuses get added we will need to add their crit buffs to here.

With the way the snapshot system was working, I ran into errors where the snapshots I would only be the last value at the end of the parse.  This was because it was being used as a reference instead of being the intended effect of being a new value each time.  I have changed it so it should work as intended now.