Added a Snapshot Module
This Snapshot module is very straight forward.  Because recommending optimal snapshotting can be very situational, we instead leave it to the user to decide if they made a good decision.

What this module does is simple, it takes in cast events of either of the DoT casts or an Iron Jaws cast, puts those into an array, and then out puts what statuses were snapshotted when they were cast formatted as a table.

This module makes use of the AdditionalStats module adding snapshot information to cast events which apply or refresh bards DoTs.