Set Bard as a supported job, and added myself as a Contributor
With the addition of a Snapshot module, and the various updates applied across the bard module reflecting the changes in Shadowbringers done in my previous update has made this bard module helpful enough to be considered supported.  While more stuff could be added, I feel that stating Bard is unsupported at this moment would give the wrong idea to users.