Fix NaN strength
When I made my Snapshot module, I changed _snapshotStatuses to copy the isActive property instead of referencing the object, because otherwise I was running into it changing during runtime causing issues.  I didn't realize that something else used this, and needed a strength property which lead to the problem here today as without it we were getting NaNs and breaking the things that relied on it.

Very simple fix when I figured it out.