Added DNC Crit buff and BRDs early DoTs
Seeing as how I am not sure if I should remove the crit tracking or not, I might as well keep it up to date.

Also, while this module is primarily for level 80 content, it can still be useful to users who are doing lower level content and wish to see how they did.  Specifically adding in checks and changes for earlier levels is a lot of work, but we can still track early abilties that get replaced for a more accurate report.