Fixed Incorrect Shadowbite and Improved Output
What started out as me just trying to fix the changelog ended with me going down a huge rabbit hole.  Now, it will properly detect how many dots were applied to the target Shadowbite was cast on.

Also, it now alerts the user of the amount of casts of Shadowbite on a single target, the amount of potency lost from each mistake, and some other small fixes and touch ups here and there.