Fixed missing downtime detection
Before, there was no check to see if the reason that a PP was "missed" was because there was downtime when WM ended. This has been rectified

Also, on the last check for a missed PP it was checking for a smaller window then other times. This also has been fixed.