Remove Scholar Ruin II Weaving Suggestion
Remove the Scholar suggestion regarding Ruin II weaving as it is no
longer accurate with the changes made to the job in Shadowbringers.

The module is detrimental to players trying to analyze their fights
currently because it overwrites the existing Weaving module that
otherwise explains poor weaving, and gives misinformation that it is
suboptimal to use Ruin II in a fight unless moving.

Signed-off-by: Panic Stevenson <Panic.Stevenson@gmail.com>