Remove Potion Module from Core
Seeing as there are existing bugs with the potions module that can lead
to misinformation, namely dying with a potion buff before 20 seconds
have passed informing the user that they used a 2.X or 3.X potion, it
makes more sense to remove the module than to leave it in for the time
being.

As per some discussion within the Discord, potion tracking may be
beyond the scope of intent for xivanalysis. Or, at the least, is not an
immediate priority.