Updates PLD suggestion text for FoF and Req
This commit fixes minor issues in suggestion text. In the FoF
module, it was still suggested to use all damaging oGCDs at least
once per window, but as Requiescat no longer deals physical damage,
it need not be used. The text has been updated to reflect this.

In the Requiescat module, when Req was used under 80% MP, the value
of the damage buff was incorrectly listed as 20% (it is now 50%), and
the other benefits of this buff were not listed. These have been changed.