Better track pre-pull draws
If we saw a play before a draw, then we had a pre-pull draw. (Theoretically we could have led with Sleeve Draw, but who does that?)
Subtract the expected remaining cooldown time from the fight duration when calculating possible future plays. If we didn't see a pre-pull draw, then we should be off cooldown.
If we're before 5.3, don't include pre-pull draws in the checklist, even if we saw one.