Added "Overcap" to the ammo timeline.
Now, the timeline will have number of rows as the maximum overcapped ammos
(which is 2 [no overcap], 3 [overcap with combo] or 4 [overcap with bloodfest]).
For each time the player had overcapped, a new entry will be placed on the timeline graph in red color
to indicate the overcapping situation, going from the "Before" entry to the "After" in a horizontal line.