BLU: Extend the Moon Flute report to support alternative actions.
Basically, all of these are mutually exclusive abilities that
share the same cooldown:

    J Kick       / Quasar
    Eruption     / Feather Rain
    Shock Strike / Mountain Buster

Using either during a moon flute is acceptable, so we should track
both.

Note that we explicitly do not track Both Ends / Nightbloom,
because Both Ends is strictly worse in almost every raiding
situation -- When are you ever going to do a fight that
has 3+ targets during every burst window?