BLU: Triple Trident and SpS need special handling
The standard burst for BLU has us using Triple Trident under
Moon Flute, but the story is a bit more complicated than that.

For Spell Speed builds, using Triple Trident on cooldown is
a (small!) DPS gain over holding it for the Moon Flute window.

This commit adds handling for this case -- If TT was off cooldown
during the burst window, either because this is the opener or
because it re-aligned itself, then the report requires it to be used.
If it was on cooldown, we don't dock them points for not being
able to use it.