BLU: Simplify the White Death cast counter
The previous version was trying to exclusively track casts that
happened when the Touch of Frost effect was active, but the way
we were adding and removing the event hook ended up mis-counting
actions in some edge cases.

So this just simplifies the entire thing, adding a hook during
initialization for the White Death action.