bugfix: Fix report loading for Firefox in Private Browsing mode
Firefox rejects Cache access in private browsing mode with security errors.
This causes reports to not load. The cache is used in two places:

* legacyStore::fetchReport
* fflogsApi::getFflogsEvents

The latter already has a try-catch block for getCache throwing an exception,
the former does not. Essentially replicates the try-catch block to the former
code location as well.

Alternatively this can be moved into fflogsApi::getCache method so that this
fix is does not need to be replicated.