BLU: New spells added in 6.45
The only change to existing spells:

    * Magic Hammer shares a cooldown with the newly-added Candy Cane
    * The Rose of Destruction & Chelonian Gate, which shared a
      cooldown, now also share it with the newly added Ruby Dynamics

Since those changes would only matter for logs generated after 6.45,
I think it's safe to just add these without messing with the layers.

NOTE: Winged Reprobation is *weird* and it's actual cooldown is
not going to be properly tracked yet.  That will be handled in
a future commit.