BLU ABC: Handle Apokalypsis
Apokalypsis is another channel like Phantom Flurry / MCH's Flamethrower.

Unlike Phantom Flurry, it does not have an optional finisher at
the end, so handling this one is even easier.

This commit both makes the ABC report count the channel as uptime,
and it adds a suggestion if people dropped out of the channel too early.