BLU: Swiftcasting Moon Flute is rarely a good idea
Moon Flute is a GCD that gives you a 15 second buff that
increases your damage by 50%; it has a 2 second cast time and
2.5 second GCD.

That means that if you hardcast it, you end up with 14.5s worth
of the buff that you can use on other GCDs -- but if you Swiftcast
it, you only end up with 12.5 seconds usable for other GCDs.

While there's some edge cases for this (you might want to swift
moon flute before downtime or before the end of the fight in order
to cram in some extra oGCDs) the general recommendation is to NOT
do this, since it'll lose you a damaging GCD under buffs.