Core Weaving module: Tighten the check
Previously this was looking at the cast time for an action and
going by a simple table to decide how many weaves you had.

The problem is that this lead to several false-negatives.

This commit makes the check slightly more strict, first by
assuming that the animation lock (for both GCDs and weaves)
is 800ms[*], then by running the equivalent of this:

    floor( (recastTime - castTime - animationLock) / animationLock )

[*] I'm told that for people with low latency this goes as low
as 600ms, but this seems a bit more forgiving for anyone not
hugging the DCs.