Core Weaving module: add a grace period for clips
Turns out that even top parses clip certain GCDs during bursts
by 20-30ms, so this adds a 50ms grace period for clips -- you're
allowed to clip a GCD by that much.

This removes a whole class of real-but-unactionable clips that the
stricter check added in the previous commit was introducing.