BLU: Fix how we track Winged Reprobation for the CD uptime metric
Essentially the same logic as before -- we only count a usage on
the 4th use of the ability -- but nwo it actually works.

Previously it was using `actors.current` to get the current amount
of Winged Reprobation buff stacks, but for whatever reason the buff
number was never correct.

Instead we just keep a simple counter.