RDM: Magic finishers under Magicked Swordplay are fine
Previously we were dinging people for this:

    `Enchanted Riposte`
        => `Enchanted Zwerchau`
            => `Enchanted Redoublement`
                => `Manafication`
                    => magic finishers
                        => second melee combo

But that's actually a pretty common way to use Manafication,
in order to fit two sets of finishers under buffs, or three
under pot.

This commir adds the magic finishers to the list of allowed
GCDs when the Magicked Swordplay buff is active.