BLU: Initial support for Dawntrail
Technically nothing has changed for BLU at 80 -- except
that we now have a bit of an inconsistency, because at
level 100, Swiftcast is 40s (down from 60s) and Addle
lasts 15s (up from 10s).

So on top of bumping support, this also overrides
those two for BLU back to their EW values.

There's still an inconsistency in the tooltips shown
in the browser (xivapi rightfully shows the values for DT)
but at least internally we do calculations using the
right values for BLU.